feat: Moved logic from NewGraphMolecule to GraphMolecule
